Watcher: Prevent triggering watch when using activate API
A wrong check in the activate watch API could lead to watches triggering
on wrong nodes. The check was supposed to check if watch execution
was distributed already in 6.x and only if not, then trigger locally.

The if-condition however was broken and triggered the watch only when
distributed watch execution was actually enabled.