HLRC: Create server agnostic request and response
The HLRC has historically reused the same Request and Response classes
that the server module uses. This commit deprecates the use of any
server module Request and Response classes, and adds a small bit of
validation logic that differs from server slightly, in that it does not
assume a check for a null ValidationException class is not enough to
determine if validation failed.

This commit very likely temporarily contains an example
Request/Response, as I do not really like the fact that a bunch of file
move's got caught up in this. I plan on splitting this up, but I wanted
to get it so people could start looking at it tomorrow.