HLRC: Replace delete watch protocol actions
This commit reverts the DeleteWatchRequest and Response move from x-pack
to protocol. It also adds new server agnostic actions to protocol. There
was some old response building in the WatcherRestHandler that also had
to be moved into a toXContent method in the old DeleteWatchResponse to
facilitate fixing the test, which is now using the server class to
serialize to XContent, and the client class to deserialize back to its
DeleteWatchRequest to ensure that the data between client and server
matches.